turtle-backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
chromadb>=0.4.24
PyPDF2>=3.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.3.1
google-generativeai>=0.3.0
